pyproject.toml
[tool.poetry]
name = "pragmatiq"
version = "0.1.0"
description = "PragmatiQ"
authors = ["Valerio Santucci <valerio.santucci.rm@gmail.com>"]

[tool.poetry.dependencies]
python = "3.12"
prometheus-client = "0.21.1"
jaeger-client = "4.8.0"
aiohttp = "3.11.14"
redis = "5.2.1"
msgpack = "1.1.0"
loguru = "0.7.3"
ray = "2.44.0"
opentelemetry-api = "^1.31.1"
opentelemetry-sdk = "^1.31.1"
opentelemetry-exporter-jaeger = "^1.21.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "8.3.5"
pytest-asyncio = "0.26.0"
pytest-mock = "3.14.0"
testcontainers = "4.9.2"
httpx = "0.28.1"
